+++ packages/bugsnag_flutter/lib/src/client.dart
@@ -17,7 +17,7 @@ import 'model.dart';
 final _notifier = {
   'name': 'Flutter Bugsnag Notifier',
   'url': 'https://github.com/bugsnag/bugsnag-flutter',
-  'version': '2.5.0'
+  'version': '3.0.0'
 };
 
 abstract class BugsnagClient {
@@ -236,7 +236,7 @@ abstract class BugsnagClient {
   void removeOnError(BugsnagOnErrorCallback onError);
 }
 
-class DelegateClient implements BugsnagClient {
+mixin DelegateClient implements BugsnagClient {
   BugsnagClient? _client;
 
   BugsnagClient get client {
@@ -340,6 +340,7 @@ class DelegateClient implements BugsnagClient {
 
 class ChannelClient implements BugsnagClient {
   FlutterExceptionHandler? _previousFlutterOnError;
+  ErrorCallback? _previousPlatformDispatcherOnError;
 
   ChannelClient(bool autoDetectErrors) {
     if (autoDetectErrors) {
@@ -347,6 +348,8 @@ class ChannelClient implements BugsnagClient {
       // same Isolate as Flutter, we install the FlutterError handler here
       _previousFlutterOnError = FlutterError.onError;
       FlutterError.onError = _onFlutterError;
+      _previousPlatformDispatcherOnError = PlatformDispatcher.instance.onError;
+      PlatformDispatcher.instance.onError = _onDispatcherError;
     }
   }
 
@@ -467,6 +470,11 @@ class ChannelClient implements BugsnagClient {
     _previousFlutterOnError?.call(details);
   }
 
+  bool _onDispatcherError(Object exception, StackTrace stack) {
+    _notifyInternal(exception, true, null, stack, null);
+    return _previousPlatformDispatcherOnError?.call(exception, stack) ?? false;
+  }
+
   Future<void> _notifyInternal(
     dynamic error,
     bool unhandled,
@@ -570,10 +578,10 @@ class ChannelClient implements BugsnagClient {
 /// is instead accessed using the [bugsnag] global:
 ///
 /// ```dart
-/// Future<void> main() => bugsnag.start(
-///   apiKey: 'your-api-key',
-///   runApp: () => runApp(MyApplication()),
-/// );
+/// Future<void> main() async {
+///   await bugsnag.start(apiKey: 'your-api-key');
+///   runApp(MyApplication());
+/// }
 /// ```
 ///
 /// See also:
@@ -587,10 +595,12 @@ class Bugsnag extends BugsnagClient with DelegateClient {
   ///
   /// Typical hybrid Flutter applications with Bugsnag will start with
   /// ```dart
-  /// Future<void> main() => bugsnag.attach(
-  ///   runApp: () => runApp(MyApplication()),
-  ///   // other configuration here
-  /// );
+  /// Future<void> main() async {
+  ///  await bugsnag.attach(
+  ///   // configuration here
+  ///   );
+  ///   runApp(MyApplication());
+  /// }
   /// ```
   ///
   /// Use this method to initialize Flutter when developing a Hybrid app,
@@ -608,14 +618,10 @@ class Bugsnag extends BugsnagClient with DelegateClient {
   /// - [addFeatureFlags]
   /// - [addOnError]
   Future<void> attach({
-    FutureOr<void> Function()? runApp,
     List<BugsnagOnErrorCallback> onError = const [],
   }) async {
-    // make sure we can use Channels before calling runApp
-    _runWithErrorDetection(
-      true,
-      () => WidgetsFlutterBinding.ensureInitialized(),
-    );
+    // make sure we can use Channels
+    WidgetsFlutterBinding.ensureInitialized();
 
     final result = await ChannelClient._channel.invokeMethod('attach', {
       'notifier': _notifier,
@@ -627,11 +633,6 @@ class Bugsnag extends BugsnagClient with DelegateClient {
     final client = ChannelClient(autoDetectErrors);
     client._onErrorCallbacks.addAll(onError);
     this.client = client;
-
-    _runWithErrorDetection(
-      autoDetectErrors,
-      () => runApp?.call(),
-    );
   }
 
   /// Initialize the Bugsnag notifier with the configuration options specified.
@@ -642,10 +643,10 @@ class Bugsnag extends BugsnagClient with DelegateClient {
   ///
   /// Typical Flutter-only applications with Bugsnag will start with:
   /// ```dart
-  /// Future<void> main() => bugsnag.start(
-  ///   apiKey: 'your-api-key',
-  ///   runApp: () => runApp(MyApplication()),
-  /// );
+  /// Future<void> main() async {
+  ///   await bugsnag.start(apiKey: 'your-api-key');
+  ///   runApp(MyApplication());
+  /// }
   /// ```
   ///
   /// See also:
@@ -657,7 +658,6 @@ class Bugsnag extends BugsnagClient with DelegateClient {
   /// - [addOnError]
   Future<void> start({
     String? apiKey,
-    FutureOr<void> Function()? runApp,
     BugsnagUser? user,
     bool persistUser = true,
     String? context,
@@ -687,23 +687,19 @@ class Bugsnag extends BugsnagClient with DelegateClient {
     Map<String, Map<String, Object>>? metadata,
     List<BugsnagFeatureFlag>? featureFlags,
     List<BugsnagOnErrorCallback> onError = const [],
-    Set<BugsnagTelemetryType>? telemetry,
+    BugsnagTelemetryTypes telemetry = BugsnagTelemetryTypes.all,
     Directory? persistenceDirectory,
     int? versionCode,
   }) async {
     final detectDartErrors =
         autoDetectErrors && enabledErrorTypes.unhandledDartExceptions;
 
-    // guarding WidgetsFlutterBinding.ensureInitialized() catches
-    // async errors within the Flutter app
-    _runWithErrorDetection(
-      detectDartErrors,
-      () => WidgetsFlutterBinding.ensureInitialized(),
-    );
-
     if (projectPackages._includeDefaults) {
       projectPackages += BugsnagProjectPackages.only(_findProjectPackages());
     }
+
+    // make sure we can use Channels
+    WidgetsFlutterBinding.ensureInitialized();
 
     await ChannelClient._channel.invokeMethod('start', <String, dynamic>{
       if (apiKey != null) 'apiKey': apiKey,
@@ -738,9 +734,7 @@ class Bugsnag extends BugsnagClient with DelegateClient {
       if (metadata != null) 'metadata': BugsnagMetadata(metadata),
       'featureFlags': featureFlags,
       'notifier': _notifier,
-      'telemetry': (telemetry ?? BugsnagTelemetryType.values)
-          .map((e) => e.toName())
-          .toList(),
+      'telemetry': telemetry,
       if (persistenceDirectory != null)
         'persistenceDirectory': persistenceDirectory.absolute.path,
       if (versionCode != null) 'versionCode': versionCode,
@@ -753,35 +747,9 @@ class Bugsnag extends BugsnagClient with DelegateClient {
     if (autoTrackSessions) {
       await resumeSession().onError((error, stackTrace) => true);
     }
-
-    _runWithErrorDetection(detectDartErrors, () => runApp?.call());
-  }
-
-  void _runWithErrorDetection(
-    bool errorDetectionEnabled,
-    FutureOr<void> Function() block,
-  ) async {
-    if (errorDetectionEnabled) {
-      await runZonedGuarded(() async {
-        await block();
-      }, _reportZonedError);
-    } else {
-      await block();
-    }
   }
 
   static const int appHangThresholdFatalOnly = 2147483647;
-
-  /// Safely report an error that occurred within a guardedZone - if attached
-  /// to a [BugsnagClient] then use its [BugsnagClient.errorHandler], otherwise push the error
-  /// upwards using [Zone.handleUncaughtError]
-  void _reportZonedError(dynamic error, StackTrace stackTrace) {
-    if (_client != null) {
-      errorHandler(error, stackTrace);
-    } else {
-      Zone.current.handleUncaughtError(error, stackTrace);
-    }
-  }
 
   static Set<String> _findProjectPackages() {
     try {